feat: CLI redesign, multi-profile support, and relational db ingestion
Overhaul the entire web-recap CLI architecture, profile detection engine,
and database ingestion layers to support rich histories without data loss:

- CLI Overhaul: Consolidated 5 complex date/time filtering parameters into
  two unified flags (`--from` / `-f` and `--to` / `-t`) supporting ISO 8601
  and relative offset strings (e.g. "today", "yesterday", "3 days").
- Subcommand Separation: Split operations into dedicated subcommands:
  - `dump` for extracting logs in multiple print formats (JSON, JSONL, CSV, Table).
  - `stats` for calculating user activity summaries and CLI ASCII charts.
  - `ingest` for copying history logs directly into external databases.
  - `list` to inspect active installations and browser profiles.
- Multi-profile harvesting: Updated browser detector to discover and harvest
  history from all user profiles automatically.
- Database Ingestion Modes: Structured `merged`, `split`, and `both` ingestion
  modes controlled by a new `--flat` flag to write either normalized relational
  schemas (using PK/FK with cascade delete) or denormalized flat data.
- Relational MongoDB Ingest: Assigned deterministic ObjectIDs based on unique
  history fields to associate parent and child documents in MongoDB bulk writes.
- Documentation: Rewrote README.md to describe new subcommands, include a Rationale
  design section, migration guide tables, and database schemas.

+++ internal/database/util_test.go
@@ -74,19 +74,19 @@ func TestConvertFirefoxTimestamp(t *testing.T) {
 func TestConvertSafariTimestamp(t *testing.T) {
 	tests := []struct {
 		name            string
-		safariVal       int64
+		safariVal       float64
 		expectZero      bool
 		minExpectedYear int
 	}{
 		{
 			name:            "Zero timestamp (Safari epoch)",
-			safariVal:       0,
+			safariVal:       0.0,
 			expectZero:      false, // 0 = 2001-01-01
 			minExpectedYear: 2001,
 		},
 		{
 			name:            "Valid timestamp",
-			safariVal:       730000000, // Some arbitrary seconds since 2001
+			safariVal:       730000000.0, // Some arbitrary seconds since 2001
 			expectZero:      false,
 			minExpectedYear: 2024, // 730M seconds since 2001 is ~2024
 		},
